requirements.txt
aiofiles==0.8.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.17.0
altgraph==0.17.4
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
packaging==24.1
pefile==2024.8.26
praw==7.7.1
prawcore==2.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==74.0.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.9.4
